requirements.txt
streamlit
torch
bitsandbytes
openai-whisper
langchain-ollama